Add files via upload
In order to utilize the application, it is essential to possess a developer app, which grants you the necessary credentials for entering them into the "token.ipynb" file. This file serves as a means to request a temporary token.

To create your application, kindly follow the provided link: https://developer.spotify.com/dashboard.

For comprehensive guidance on developing and utilizing the application, please refer to the following references:

Documentation on the web API: https://developer.spotify.com/documentation/web-api.